% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.dfonly.R
\name{is.dfonly}
\alias{is.dfonly}
\title{Is it a data frame and only a data frame?}
\usage{
is.dfonly(x)
}
\arguments{
\item{x}{an object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Is it a data frame and only a data frame?
}
\examples{
x <- as.data.frame(matrix(1:4,2,2))
is.dfonly(x)

class(x) <- c("tbl_df", "tbl", "data.frame")
x
is.dfonly(x)

}
