% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.path.R
\name{is.path}
\alias{is.path}
\title{Is it a path?}
\usage{
is.path(x)
}
\arguments{
\item{x}{an object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Is it a path?
}
\examples{
is.path(as.character(NA))
is.path(as.character(NULL))
is.path(tempdir())
is.path(file.path(tempdir(),"uno"))

}
