% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.string.R
\name{is.string}
\alias{is.string}
\title{Is it a non-NA single string?}
\usage{
is.string(x)
}
\arguments{
\item{x}{an object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Is it a non-NA single string?
}
\examples{
is.string(as.character(NA))
is.string(as.character(NULL))
is.string("a")
is.string(c("a","b"))
is.string("b")

}
