% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.TF.R
\name{is.TF}
\alias{is.TF}
\title{Is it TRUE or FALSE?}
\usage{
is.TF(x)
}
\arguments{
\item{x}{an object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Is it TRUE or FALSE?
}
\examples{
is.TF(as.logical(NA))
is.TF(as.logical(NULL))
is.TF(TRUE)
is.TF(c(TRUE,TRUE))
is.TF(1)

}
